+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 2025-09-15
 
 - Fix caching mechanism of the latest block's hash [#4479](https://github.com/lambdaclass/ethrex/pull/4479)
+- Add `jemalloc` as an optional global allocator used by default [#4301](https://github.com/lambdaclass/ethrex/pull/4301)
 
 ### 2025-09-11